requirements.txt
tensorflow
python-box==6.0.2
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
boto3
python-dotenv
-e .